services/background-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
loguru==0.7.2
pydantic==2.5.0
pillow==10.1.0
numpy==1.25.2
opencv-python==4.8.1.78
rembg==2.0.53
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
diffusers==0.24.0
transformers==4.36.2
accelerate==0.25.0
safetensors==0.4.1
requests==2.31.0
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2